Travis CI: build with HarfBuzz
On OS X, disable some unnecessary HarfBuzz dependencies. This triggers
a source build of HarfBuzz, but it should be fast and avoids bringing
in Fontconfig through a dependency chain, which we want to avoid as it
wastes a lot of time building its cache when installed.

The dependency that brings in Fontconfig is gobject-introspection, but
we don't need icu4c either, so disable that to save a little more time
that would be spent installing icu4c. We could also disable glib, but
the fribidi formula also has it as a dependency and brings it in anyway.