[RFC] \box tag
This essentially sets the margins. But it allows you to set different
margins for top and bottom (instead of only a "MarginV"), and takes a
rectangle instead of right/bottom margin values. Rather similar to for
example \clip.

Example: \box(500,50,700,100)

For now, either the top or bottom margin is ignored, depending on
alignment. A future idea might be to cut off the text in a non-ugly way,
such as ellipsization. (This would be controlled by other tags.)

This tag should be a nice addition in contexts where setting the the
event or style margins is too bothersome.