api: add ass_track_set_feature()
Since C does not allow empty enums, there is an "example", which doesn't
do anything. I think we might be able to make this change the default
bidi direction or so. As if this commit, the flag set by it is also not
available outside of ass.c, which should be solved by moving parser_priv
to an internal header.