ass_parse: improve performance of tag name comparisons
Yes, this actually makes a major difference on some tag-heavy scripts.

This lets the whole function get inlined, rather than making a call out
to the libc's strcmp implementation. The libc's version is likely
much faster on longer strings, but we're only ever comparing against
strings that are a few characters long, so that doesn't really matter.
It also avoids making an extra pass for the strlen.