Ignore font substitution when splitting combined bitmaps
Well, that was easy.

font->desc.family is the raw family name from FontName/\fn,
barring the leading @ for vertical fonts. So, for reference,
even if the script requests two different font names and neither
exists on the system and font fallback picks the same font
for both, is_new_bm_run will treat them as distinct.
This is indeed the desired behaviour that matches VSFilter.