Introduce alloc wrappers
This allos detailed logging of allocation failures and can in the future
be expanded to alow API-users to supply libass with a custom allocator
API, which may be useful on some devices or in fuzzing to simulate
allocation failures.

This moves the strndup replacment from ass_utils to ass_alloc and also
adds a strdup replacement as neither of them is part of ISO C99, but
POSIX and the upcoming ISO C2x.

TODO:
 Actually replacing every allocation in libass with these wrappers