coretext: return full/PostScript name on fallback
CoreText's "family name" can use TT_NAME_ID_TYPOGRAPHIC_FAMILY names,
which we ignore for compatibility with VSFilter. As a result, fallback
returns a family name that we then fail to find, and fallback fails.

Return a fuller name (fullname or postscriptname as appropriate)
instead, which should be more reliable.