WHOLE_TEXT: give HarfBuzz context to shape correctly across runs
When WHOLE_TEXT is disabled, keep calling HarfBuzz with no context at all.
It might seem sensible to give it the whole VSFilter-run as context so
it could shape Arabic correctly across font fallback in case of weird fonts,
but it seems that Uniscribe doesn't behave like this in VSFilter.

For testing that, I created a font with only two glyphs: an isolated ain
and a medial ain; and a GSUB medi lookup that links them together.
Setting \fnMyTestFont on an Arabic word with a medial ain, libass
with WHOLE_TEXT (i. e. HarfBuzz with context) showed the medial shape,
but VSFilter (i. e. Uniscribe) showed the isolated shape.
BorderStyle 3 confirmed that VSFilter was treating the string
as a single run and therefore passing it to GDI/Uniscribe as a whole.