ass: process format lines only once
This speeds up parsing and removes strdup's as a failure source.
For ABI and API compatibility we keep the format strings in the track
struct around and also store a copy internally so we can detect changes
made by API users via direct struct access.
We may want to consider restricting this direct access in the future.