configure: Replace obsolete macros
AC_CONFIG_HEADER has been replaced with AC_CONFIG_HEADERS for a long
time and recent version of autotools started complaining about it.

AC_HEADER_STDC has been deemed "obsolescent" and unneeded for
"new programs" since at least autotools 2.60, which was released 2006.
On top of it we're not even using its output variable anywhere to
check if the system headers do in fact conform to the C standard, so
remove this superfluous test.

AC_HEADER_STDBOOL is not obsolecent, but simalarly to AC_HEADER_STDC we
never use its output anywhere, so it can be removed as well.