ass_feature: ass_track: Use a bitmask for features
Previously we use a fallthrough, which makes the logic pretty complex
since things either need to be inclusive (in the case of
ASS_FEATURE_INCOMPATIBLE_EXTENSIONS) or exclusive (in the case of
selecting an individual feature).

Putting these in a private bitmask means we can trivially select which
to do at runtime without much extra code complexity, and we also can
change tack later if needs be.