Make fix_outline more correct in presence of blur
Blurred alphas are not coverage values. The area that is not
covered by the original glyph is a mix of the border and the
background, governed by the blurred border opacities. To get
an exact result, we would need to integrate the blurred
opacities on fragments of pixels that exclude the original
glyph (or, equivalently, on fragments that are covered by it,
as these fragments are complementary). This is too hard,
and in fact blur itself already employs approximations
to avoid integration even on whole pixels.

Give up on exact values when blur is applied, and treat
each pixel of the blurred border image as a uniformly flat
translucent square.

Note that we can in this case get non-zero opacities even when
the glyph's coverage is larger than the border's uncorrected
opacity, invalidating our current o[x] > g[x] sanity check.

Writing and solving equations like before, we want:
         a_g  (A_g c_g + (1 - A_g) c_b) +
    (1 - a_g) ((A_o a_o) c_o + (1 - A_o a_o) c_b)

and we obtain:
       A_o (c_o - c_b) = 0 and a_o' does not matter
    or A_g a_g = 1 and a_o' does not matter
    or a_o' = (a_o - a_g a_o) / (1 - A_g a_g)

Handle the first case like the third,
and handle the second case by setting a_o' = 0.

The observable difference for non-blurred bitmaps
was not large in my tests, but the code is simple,
so the exact non-blurred calculations are retained.