ass_render: Break at last position if no space was seen
It's pretty common (in Chinese/Japanese/Korean especially) to not have
any spaces to split on. Currently, these cases simply overflow the
render area.

If this happens, try a last ditch attempt to wrap at the last character
seen. This dramatically improves the usability of long subtitles with no
clear break point.

Compared to https://github.com/libass/libass/pull/372, which is about
Unicode line breaking, this pull request is much simpler (but
complementary): we just split in any case where we will overflow the
buffer as a last resort.

Example previous issues which show demand for this behaviour:

- https://github.com/mpv-player/mpv/issues/8115
- https://github.com/mpv-player/mpv/issues/2203
- https://github.com/libass/libass/issues/348
- https://github.com/libass/libass/issues/360