Make fix_outline always produce the correct final color (bar blur)
Whenever a glyph has a border, the original glyph's own ASS_Image
is applied on top of the border's ASS_Image. Let's mathematically
devise such alpha values for the border image that the resulting
colors exactly match (modulo rounding) the colors produced by a
hypothetical ASS_Image that contains both the glyph and the border.

Let a_i be the opacity of a single pixel within an ASS_Image,
where i is either g for glyph or o for outline (border). Further,
let A_i be the opacity and c_i the color of the entire ASS_Image.

This commit only considers cases when no blur is applied,
so that the border is a solid, expanded copy of the glyph.
Blur will be addressed in a separate commit.

Think of a_i as the _area_ covered by the glyph/border,
measured in pixels. We have 0 <= a_g <= a_o <= 1.
We also have 0 <= A_g, A_o <= 1.

We need a background color. Let's denote it as c_b. The background
color includes any effects from applying the shadow of the glyph.

The final color we want to achieve is:
  * on the area covered by the original glyph,
                                       A_g c_g + (1 - A_g) c_b.
  * on the area covered by the border, A_o c_o + (1 - A_o) c_b.
  * on the remaining area,             c_b.
Averaging the color over the pixel's area gives:
           a_g  (A_g c_g + (1 - A_g) c_b) +
    (a_o - a_g) (A_o c_o + (1 - A_o) c_b) +
      (1 - a_o)  c_b

The color we actually get after blending the border with the
background and then blending the original glyph with the result is:
        (A_g a_g) c_g +
    (1 - A_g a_g) ((A_o a_o) c_o + (1 - A_o a_o) c_b)

We want to replace a_o in the actual color's expression with a
value that makes it match the expected color. Let's call this value
a_o'. Equating both expressions and solving for a_o' gives:
       A_o (c_o - c_b) = 0 and a_o' does not matter
    or A_g a_g = 1 and a_o' does not matter
    or a_o' = (a_o - a_g) / (1 - A_g a_g)

Use the solution formula from the third case to calculate a_o',
as the first two cases allow a_o' to be anything.

The calculation itself is performed and rounded (half towards
+infinity) using floating-point division because this is
actually faster than integer division and because this is
vectorizable on x86 (massively improving the speed further).

The only issue is if the 1 - A_g a_g denominator should be zero.
A_g a_g can only equal 1 if A_g = a_g = 1. Since a_g <= a_o <= 1,
a_o must also be 1. As it happens, libass already handles this case
separately: as a precaution, it checks that a_g < a_o and sets a_o'
to zero if this condition fails to hold. Retain this check for the
numerator, still as a precaution. Unfortunately, using this check
for the denominator or the whole formula seems to break automatic
vectorization of the loop, so use a trick: instead of 1 - A_g a_g,
use nextafter(1) - A_g a_g and accept the very rare and minor
precision loss. This is just enough to avoid division by zero
and allows the loop to be autovectorized. Since in this case a_o'
can be any value, this does not adversely affect our equation
solving. Moreover, if A_g a_g = 1, then the numerator must be 0,
so we will, in fact, produce the same value (0) as we do now with
the a_g < a_o check.

The formula requires knowledge of A_g. To enable this, the opacity of
the original glyph is passed to fix_outline and (via CompositeHashKey)
to ass_composite_construct. Generated composite bitmaps depend on it,
so composites with different primary/secondary color alpha (depending
on karaoke) are now distinguished by the composite cache.

NB: this commit produces phantom, fully transparent ASS_Images
that duplicate parts of border images when \kf is in effect.
This may be a problem if any consumer assigns any meaning
to fully transparent ASS_Images.