Bidi: remember resolved base direction
There used to be a FIXME comment about this,
but it was removed in commit 27cc03363cfbddb9877cd547327d56405653add7.

Reordering works fine without this *most of the time*, but this is still
a bug. It affects one thing only: trailing whitespace in each line/run.

Of course, we remove trailing U+0020 SPACE characters from each line
regardless of bidi, but Unicode bidi recognizes other whitespace,
and we can now also have trailing U+0020 within in-line runs.

With right-to-left base direction, all trailing whitespace within
each line (after line wrapping) or run are to be placed at the leftmost
end in visual order (even if that whitespace is nominally part
of a left-to-right embed). Conversely, with our current code that
always tells FriBidi the base direction is left-to-right here,
trailing whitespace is always placed at the rightmost end of each line,
even if it is inside right-to-left text (which makes it appear
as *leading* whitespace to the reader).

We treat explicit line breaks \N as paragraph separators, where each
paragraph can have its own base direction. (For what it's worth, this
is desirable for WebVTT. This also matches a common behavior of plain
text editors.) Therefore, we must remember each paragraph's resolved
direction separately.