wrap_lines: do not merge linebreaks
Merging linebreaks will always result in text overflowing its designated
bounds, but integer overflows or other bugs may erroneously make it
seems like it would be beneficial. If this is not prevented it can also
lead to out-of-bound reads as in CVE-2016-7969.

This replaces and reverts f4f48950788b91c6a30029cc28a240b834713ea7
which did not prevent linebreak-merges and turned
hard linebreaks into soft ones on merge.