coretext: deduplicate found fonts in match_fonts
We ask Core Text to match multiple name fields, and a single font may
match several at once. In this case, by default, Core Text duly returns
the font multiple times, and we add it to fontselect multiple times.
This does not break anything, but it is a waste, so ask Core Text
to deduplicate the returned list.

This is supposedly available since Mac OS X 10.5.
We support Core Text on 10.6+.