ass_shaper_find_runs: don't resolve default-ignorable glyphs
default-ignorable (DI) codepoints glyphs are not present at all fonts,
are not used by the simple shaper for rendering, and are mostly not
used by the complex shaper for rendering.

This avoids missing-glyph warnings for DI codepoints.

Because now we don't try to resolve DI glyphs - their face_index could
differ from their surrounding, which would break the run. Similarly,
before this commit DI could also break the run due to missing glyphs
or if resolved-with-different-face.

So now we also don't allow DI face_index to break the run.

However, DI face_index can still differ from the surrounding, so we
adjust DI face_index to match the rest of the run if possible so that
the complex shaper later gets the correct face_index if needed
(technically, only the first face_index of a run is used later).

There's still a potential issue if the complex shaper needs DI glyphs
for the rendering but they're not available at the run's face_index.

After discussion with @astiob, it was concluded that perfect solution
can't exist without major refactoring of the dataflow, and that the
the approach at this commit (don't resolve DI glyphs, don't break runs
on DI, adjust DI face_index to match the run) is both better than
before and also the best we can do within the current code form.