ass: treat 'Name' and 'Actor' as synonyms in format lines
Commit b3339152db5b80991966911fe2efc8a6174b0071 explicitly defaulted
ScaledBorderAndShadow to "yes" for files with custom format lines to
allow a subsequent commit to change the global default to "no" to match
VSFilter, while trying to avoid breaking files expecting libass' prior
default. This commit also changed the fallback ASS event format line to
use 'Name' instead of 'Actor' to match what we are actually parsing,
what Aegisub is emitting and what is used in the "specification".

However, as it turns out, PopSub and files _converted_ by VSFilter
actually do use 'Actor', but are expecting the VSFilter-compatible
default "no".
Thus, treat both variants as synonyms in event parsing and format line
comparison. There is a slight inconsistency here, that _Style_ format
lines which use _Actor_ instead of _Name_ are not recognised as a custom
format line — but a Style Format without a 'Name' won't be able to be
used in rendering anyway.

fixes https://github.com/libass/libass/issues/516