configure: drop disused Automake conditional for 32-bit x86
It was used when we had ASFLAGS logic in Makefile.am,
but all ASFLAGS logic has been consolidated in configure.ac
and we have not used this conditional ever since.
Unlike the conditional flagging x86-64, which can be used
for building x86-64-only assembler files, it does not seem
that the 32-bit conditional will ever be useful again.