fontselect: don't leak substitution fullname array if it's empty
Discovered by OSS-Fuzz.

This is a longstanding bug (since 0.13.0 released in 2015, when
the new fontselect was first introduced), but it should be rare:
it only manifests when Fontconfig has no substitutions configured
(which is unusual) or when font name strdups fail.

Fixes https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=39257.