x86: fix AVX detection
The removed check compared some bits from the vendor string to the ones
from "GenuineIntel". This blocked AVX and AVX2 from being used on
AMD- and other non-Intel-CPUs and even on Intel-CPUs if run in a
hypervisor or VM which changes the vendor string (eg " KVMKVMKVM  ").

Drop the vendor check to utilise AVX on all supporting CPUs.