ass.h: fix compilation with old gcc
While the deprecated attribute wass added to GCC in
https://gcc.gnu.org/git/?p=gcc.git;a=commit;h=e23bd2185fae41772d012231e005b789b3a9e6dc
first released in GCC 3.1.0, support for an argument was only added in
https://gcc.gnu.org/git/?p=gcc.git;a=commit;h=9b86d6bb25587db93a322bf5778e9892aaa8b776
first released in GCC 4.5.0.
Since we did not check for argument support before,
building with a GCC between the above two changes was broken.

The commit assumes Clang always supported an argument,
but that has not actually been verified.

Reported in https://github.com/libass/libass/issues/595